cleanup: Got rid of a bunch of DWORD/WORD stuff in guilib primarily - colors, alignment, fonts, labels, keys and actions all purged. cleanup: Removed <windows.h> from StdString.h, and removed system.h from a bunch of headers in guilib.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@23060 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/guilib/common/SDLJoystick.h
@@ -66,7 +66,7 @@ private:
   int m_SafeRange; // dead zone
   Uint32 m_pressTicksButton;
   Uint32 m_pressTicksHat;
-  WORD m_ActiveFlags;
+  Uint8 m_ActiveFlags;
   std::vector<SDL_Joystick*> m_Joysticks;
   std::vector<std::string> m_JoystickNames;
 };
